refactor: move fetchGitHubAPI method to core
